+++ src/__tests__/FixedSizeGrid.js
@@ -155,7 +155,7 @@ describe('FixedSizeGrid', () => {
       <FixedSizeGrid {...defaultProps} />
     );
     const scrollContainer = findScrollContainer(rendered);
-    expect(scrollContainer.props.style.pointerEvents).toBe('');
+    expect(scrollContainer.props.style.pointerEvents).toBe(undefined);
     rendered.getInstance().setState({ isScrolling: true });
     expect(scrollContainer.props.style.pointerEvents).toBe('none');
   });
@@ -431,31 +431,30 @@ describe('FixedSizeGrid', () => {
         })
       );
     });
+
+    it('should ignore offsets less than zero', () => {
+      const onScroll = jest.fn();
+      const instance = ReactDOM.render(
+        <FixedSizeGrid {...defaultProps} onScroll={onScroll} />,
+        document.createElement('div')
+      );
+      instance.scrollTo({ scrollLeft: 100, scrollTop: 100 });
+      onScroll.mockClear();
+      instance.scrollTo({ scrollLeft: -1, scrollTop: -1 });
+      expect(onScroll.mock.calls[0][0].scrollLeft).toBe(0);
+      expect(onScroll.mock.calls[0][0].scrollTop).toBe(0);
+    });
   });
 
   describe('scrollToItem method', () => {
     it('should not set invalid offsets when the list contains few items', () => {
-      const onScroll = jest.fn();
       const rendered = ReactTestRenderer.create(
-        <FixedSizeGrid
-          {...defaultProps}
-          columnCount={1}
-          rowCount={2}
-          onScroll={onScroll}
-        />
+        <FixedSizeGrid {...defaultProps} columnCount={1} rowCount={2} />
       );
-      onScroll.mockClear();
-      // Offset should not be negative.
-      rendered
-        .getInstance()
-        .scrollToItem({ columnIndex: 0, rowIndex: 0, align: 'auto' });
-      expect(onScroll).toHaveBeenCalledWith({
-        horizontalScrollDirection: 'backward',
-        scrollLeft: 0,
-        scrollTop: 0,
-        scrollUpdateWasRequested: true,
-        verticalScrollDirection: 'backward',
-      });
+      expect(onItemsRendered).toMatchSnapshot();
+      onItemsRendered.mockClear();
+      rendered.getInstance().scrollToItem(0);
+      expect(onItemsRendered).not.toHaveBeenCalled();
     });
 
     it('should scroll to the correct item for align = "auto"', () => {
@@ -563,6 +562,41 @@ describe('FixedSizeGrid', () => {
       rendered.getInstance().scrollToItem({ rowIndex: 10, align: 'center' });
       // Scroll left to column 3, without changing scrollTop
       rendered.getInstance().scrollToItem({ columnIndex: 3, align: 'center' });
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
+    });
+
+    it('should scroll to the correct item for align = "smart"', () => {
+      const rendered = ReactTestRenderer.create(
+        <FixedSizeGrid {...defaultProps} />
+      );
+
+      // Scroll down enough to show item 10 at the center.
+      // It was further than one screen away, so it gets centered.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 10, rowIndex: 10, align: 'smart' });
+      // No need to scroll again; item 9 is already visible.
+      // Overscan indices will change though, since direction changes.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 9, rowIndex: 9, align: 'smart' });
+      // Scroll up enough to show item 2 as close to the center as we can.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 2, rowIndex: 2, align: 'smart' });
+      // Scroll down to row 10, without changing scrollLeft
+      rendered.getInstance().scrollToItem({ rowIndex: 10, align: 'smart' });
+      // Scroll left to column 0, without changing scrollTop
+      rendered.getInstance().scrollToItem({ columnIndex: 0, align: 'smart' });
+
+      // Scrolling within a distance of a single screen from viewport
+      // should have the 'auto' behavior of scrolling as little as possible.
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 5, rowIndex: 5, align: 'smart' });
+      rendered
+        .getInstance()
+        .scrollToItem({ columnIndex: 10, rowIndex: 10, align: 'smart' });
       expect(onItemsRendered.mock.calls).toMatchSnapshot();
     });
 
@@ -678,6 +712,30 @@ describe('FixedSizeGrid', () => {
         scrollUpdateWasRequested: true,
         verticalScrollDirection: 'backward',
       });
+    });
+
+    it('should ignore indexes less than zero', () => {
+      const instance = ReactDOM.render(
+        <FixedSizeGrid {...defaultProps} />,
+        document.createElement('div')
+      );
+      instance.scrollToItem({ columnIndex: 20, rowIndex: 20 });
+      onItemsRendered.mockClear();
+      instance.scrollToItem({ columnIndex: -1, rowIndex: -1 });
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
+    });
+
+    it('should ignore indexes greater than itemCount', () => {
+      const instance = ReactDOM.render(
+        <FixedSizeGrid {...defaultProps} />,
+        document.createElement('div')
+      );
+      onItemsRendered.mockClear();
+      instance.scrollToItem({
+        columnIndex: defaultProps.columnCount * 2,
+        rowIndex: defaultProps.rowCount * 2,
+      });
+      expect(onItemsRendered.mock.calls).toMatchSnapshot();
     });
   });